Refactor referral detail tabs and extract utilities
Split ReferralDetail page into modular tab components (ReferralDetailsTab, ReferrerIndividualTab, AuditHistoryTab) and move formatting logic to new utils/formatters module. Centralize referral-related labels in constants/referralLabels. Update imports and exports accordingly for improved maintainability and code organization.
