+++ admin-app/src/components/index.ts
@@ -1,2 +1,3 @@
 export { Layout } from "./Layout";
 export * from "./ui";
+export * from "./referral";
