+++ admin-app/src/components/referral/index.ts
@@ -1,0 +1,3 @@
+export { ReferralDetailsTab } from "./ReferralDetailsTab";
+export { ReferrerIndividualTab } from "./ReferrerIndividualTab";
+export { AuditHistoryTab } from "./AuditHistoryTab";
