+++ admin-app/src/constants/index.ts
@@ -1,1 +1,2 @@
 export * from "./labels";
+export * from "./referralLabels";
