+++ admin-app/src/constants/referralLabels.ts
@@ -1,0 +1,52 @@
+export const yesNoUnknownLabels: Record<string, string> = {
+  YES: "Yes",
+  NO: "No",
+  UNKNOWN: "Unknown",
+};
+
+export const releaseFromLabels: Record<string, string> = {
+  NO: "No",
+  HOSPITAL_MEDICAL_FACILITY: "Hospital/Medical Facility",
+  CORRECTIONS: "Corrections",
+  YOUTH_TRANSITION_MCFD: "Youth Transition (MCFD)",
+  YOUTH_TRANSITION_DELEGATED_ABORIGINAL_AGENCY:
+    "Youth Transition (Delegated Aboriginal Agency)",
+  ALCOHOL_DRUG_FACILITY: "Alcohol/Drug Facility",
+};
+
+export const supportLabels: Record<string, string> = {
+  CULTURAL: "Cultural",
+  COMMUNITY_SUPPORTS: "Community Supports",
+  FOOD_SECURITY: "Food Security",
+  HOUSING: "Housing",
+  INCOME_ASSISTANCE_PROVINCIAL: "Income Assistance (Provincial)",
+  INCOME_ASSISTANCE_FEDERAL: "Income Assistance (Federal)",
+  MENTAL_HEALTH: "Mental Health",
+  SYSTEM_NAVIGATION: "System Navigation",
+  HEALTH_SERVICES: "Health Services",
+  SUBSTANCE_USE: "Substance Use",
+  INDIGENOUS_SUPPORTS: "Indigenous Supports",
+  INTEGRATED_JUSTICE_SUPPORTS: "Integrated Justice Supports",
+  OTHERS: "Others",
+};
+
+export const statusLabels: Record<string, string> = {
+  OPEN: "Open",
+  ASSIGNED: "Assigned",
+  CONTACT_MADE: "Contact Made",
+  CLOSED: "Closed",
+};
+
+export const outcomeLabels: Record<string, string> = {
+  BCEA_APPLICATION_SUBMITTED: "BCEA Application Submitted",
+  BCEA_APPLICATION_COMPLETED_FILE_OPENED:
+    "BCEA Application Completed - File Opened",
+  SUPPLEMENTS_ISSUED: "Supplements Issued",
+  CASE_MANAGED: "Case Managed",
+  SERVICES_PROVIDED: "Services Provided",
+  NOT_LOCATED: "Not Located",
+  LOCATED_REFUSED_SERVICE: "Located - Refused Service",
+  NON_APPROPRIATE_REFERRAL_RETURNED: "Non-Appropriate Referral - Returned",
+  REFERRED_TO_VS_CS: "Referred to VS/CS",
+  REFERRED_TO_COMMUNITY_PARTNER: "Referred to Community Partner",
+};
